v1.2.9

+++ upload/admin/controller/extension/feed.php
@@ -14,7 +14,7 @@ class ControllerExtensionFeed extends Controller {
    		);
 
    		$this->document->breadcrumbs[] = array(
-       		'href'      => $this->url->https('extention/feed'),
+       		'href'      => $this->url->https('extension/feed'),
        		'text'      => $this->language->get('heading_title'),
       		'separator' => ' :: '
    		);
@@ -24,13 +24,16 @@ class ControllerExtensionFeed extends Controller {
 		$this->data['text_no_results'] = $this->language->get('text_no_results');
 
 		$this->data['column_name'] = $this->language->get('column_name');
-		$this->data['column_development'] = $this->language->get('column_development');
 		$this->data['column_status'] = $this->language->get('column_status');
 		$this->data['column_action'] = $this->language->get('column_action');
 
 		$this->data['success'] = @$this->session->data['success'];
 		
 		unset($this->session->data['success']);
+
+    	$this->data['error'] = @$this->session->data['error'];
+    
+		unset($this->session->data['error']);
 
 		$this->load->model('setting/extension');
 
@@ -65,10 +68,9 @@ class ControllerExtensionFeed extends Controller {
 			}
 									
 			$this->data['extensions'][] = array(
-				'name'        => $this->language->get('heading_title'),
-				'development' => $this->language->get('text_development'),
-				'status'      => $this->config->get($extension . '_status') ? $this->language->get('text_enabled') : $this->language->get('text_disabled'),
-				'action'      => $action
+				'name'   => $this->language->get('heading_title'),
+				'status' => $this->config->get($extension . '_status') ? $this->language->get('text_enabled') : $this->language->get('text_disabled'),
+				'action' => $action
 			);
 		}
 						
@@ -80,27 +82,39 @@ class ControllerExtensionFeed extends Controller {
 	}
 	
 	public function install() {
-		$this->load->model('setting/extension');
+    	if (!$this->user->hasPermission('modify', 'extension/feed')) {
+      		$this->session['error'] = $this->language->get('error_permission'); 
+			
+			$this->redirect($this->url->https('extension/feed'));
+    	} else {
+			$this->load->model('setting/extension');
 		
-		$this->model_setting_extension->install('feed', $this->request->get['extension']);
+			$this->model_setting_extension->install('feed', $this->request->get['extension']);
 		
-		$this->load->model('user/user_group');
+			$this->load->model('user/user_group');
 		
-		$this->model_user_user_group->addPermission($this->user->getId(), 'access', 'feed/' . $this->request->get['extension']);
-		$this->model_user_user_group->addPermission($this->user->getId(), 'modify', 'feed/' . $this->request->get['extension']);
+			$this->model_user_user_group->addPermission($this->user->getId(), 'access', 'feed/' . $this->request->get['extension']);
+			$this->model_user_user_group->addPermission($this->user->getId(), 'modify', 'feed/' . $this->request->get['extension']);
 		
-		$this->redirect($this->url->https('extension/feed'));
+			$this->redirect($this->url->https('extension/feed'));			
+		}
 	}
 	
 	public function uninstall() {
-		$this->load->model('setting/extension');
-		$this->load->model('setting/setting');
+    	if (!$this->user->hasPermission('modify', 'extension/feed')) {
+      		$this->session['error'] = $this->language->get('error_permission'); 
+			
+			$this->redirect($this->url->https('extension/feed'));
+    	} else {		
+			$this->load->model('setting/extension');
+			$this->load->model('setting/setting');
 		
-		$this->model_setting_extension->uninstall('feed', $this->request->get['extension']);
+			$this->model_setting_extension->uninstall('feed', $this->request->get['extension']);
 		
-		$this->model_setting_setting->deleteSetting($this->request->get['extension']);
+			$this->model_setting_setting->deleteSetting($this->request->get['extension']);
 		
-		$this->redirect($this->url->https('extension/feed'));	
+			$this->redirect($this->url->https('extension/feed'));
+		}
 	}
 }
 ?>
