+++ upload/admin/controller/extension/shipping.php
@@ -14,7 +14,7 @@ class ControllerExtensionShipping extends Controller {
    		);
 
    		$this->document->breadcrumbs[] = array(
-       		'href'      => $this->url->https('extention/shipping'),
+       		'href'      => $this->url->https('extension/shipping'),
        		'text'      => $this->language->get('heading_title'),
       		'separator' => ' :: '
    		);		
@@ -24,7 +24,6 @@ class ControllerExtensionShipping extends Controller {
 		$this->data['text_no_results'] = $this->language->get('text_no_results');
 				
 		$this->data['column_name'] = $this->language->get('column_name');
-		$this->data['column_development'] = $this->language->get('column_development');
 		$this->data['column_status'] = $this->language->get('column_status');
 		$this->data['column_sort_order'] = $this->language->get('column_sort_order');
 		$this->data['column_action'] = $this->language->get('column_action');
@@ -32,6 +31,10 @@ class ControllerExtensionShipping extends Controller {
 		$this->data['success'] = @$this->session->data['success'];
 		
 		unset($this->session->data['success']);
+		
+    	$this->data['error'] = @$this->session->data['error'];
+    
+		unset($this->session->data['error']);
 
 		$this->load->model('setting/extension');
 
@@ -66,11 +69,10 @@ class ControllerExtensionShipping extends Controller {
 			}
 									
 			$this->data['extensions'][] = array(
-				'name'        => $this->language->get('heading_title'),
-				'development' => $this->language->get('text_development'),
-				'status'      => $this->config->get($extension . '_status') ? $this->language->get('text_enabled') : $this->language->get('text_disabled'),
-				'sort_order'  => $this->config->get($extension . '_sort_order'),
-				'action'      => $action
+				'name'       => $this->language->get('heading_title'),
+				'status'     => $this->config->get($extension . '_status') ? $this->language->get('text_enabled') : $this->language->get('text_disabled'),
+				'sort_order' => $this->config->get($extension . '_sort_order'),
+				'action'     => $action
 			);
 		}
 	
@@ -82,27 +84,39 @@ class ControllerExtensionShipping extends Controller {
 	}
 	
 	public function install() {
-		$this->load->model('setting/extension');
+		if (!$this->user->hasPermission('modify', 'extension/shipping')) {
+			$this->session['error'] = $this->language->get('error_permission'); 
+			
+			$this->redirect($this->url->https('extension/shipping'));
+		} else {		
+			$this->load->model('setting/extension');
 		
-		$this->model_setting_extension->install('shipping', $this->request->get['extension']);
+			$this->model_setting_extension->install('shipping', $this->request->get['extension']);
 
-		$this->load->model('user/user_group');
+			$this->load->model('user/user_group');
 		
-		$this->model_user_user_group->addPermission($this->user->getId(), 'access', 'shipping/' . $this->request->get['extension']);
-		$this->model_user_user_group->addPermission($this->user->getId(), 'modify', 'shipping/' . $this->request->get['extension']);
+			$this->model_user_user_group->addPermission($this->user->getId(), 'access', 'shipping/' . $this->request->get['extension']);
+			$this->model_user_user_group->addPermission($this->user->getId(), 'modify', 'shipping/' . $this->request->get['extension']);
 
-		$this->redirect($this->url->https('extension/shipping'));
+			$this->redirect($this->url->https('extension/shipping'));
+		}
 	}
 	
 	public function uninstall() {
-		$this->load->model('setting/extension');
-		$this->load->model('setting/setting');
+		if (!$this->user->hasPermission('modify', 'extension/shipping')) {
+			$this->session['error'] = $this->language->get('error_permission'); 
+			
+			$this->redirect($this->url->https('extension/shipping'));
+		} else {		
+			$this->load->model('setting/extension');
+			$this->load->model('setting/setting');
 		
-		$this->model_setting_extension->uninstall('shipping', $this->request->get['extension']);
+			$this->model_setting_extension->uninstall('shipping', $this->request->get['extension']);
 		
-		$this->model_setting_setting->deleteSetting($this->request->get['extension']);
+			$this->model_setting_setting->deleteSetting($this->request->get['extension']);
 		
-		$this->redirect($this->url->https('extension/shipping'));	
+			$this->redirect($this->url->https('extension/shipping'));
+		}
 	}
 }
 ?>
